+++ starsky/starsky/clientapp/src/components/organisms/preferences-publish-profiles/internal/editable-profile-item.ts
@@ -1,0 +1,6 @@
+import { IAppSettingsPublishProfileItem } from "../../../../interfaces/IAppSettingsPublishProfiles";
+
+export interface EditableProfileItem extends IAppSettingsPublishProfileItem {
+  _id: string;
+  optimizersText: string;
+}
