+++ starsky/starsky/clientapp/src/components/organisms/preferences-publish-profiles/internal/editable-profile.ts
@@ -1,0 +1,6 @@
+import { EditableProfileItem } from "./editable-profile-item";
+
+export interface EditableProfile {
+  name: string;
+  items: EditableProfileItem[];
+}
